requirements.txt
altgraph==0.16.1
astroid==2.2.5
beautifulsoup4==4.7.1
et-xmlfile==1.0.1
future==0.17.1
gevent==1.4.0
greenlet==0.4.15
isort==4.3.16
jdcal==1.4
lazy-object-proxy==1.3.1
macholib==1.11
mccabe==0.6.1
msgpack-python==0.5.6
openpyxl==2.6.1
pefile==2018.8.8
PyInstaller==3.4
pylint==2.3.1
PyYAML==5.1
pyzmq==18.0.1
six==1.12.0
soupsieve==1.9
typed-ast==1.3.1
wrapt==1.11.1
zerorpc==0.6.1
